Finally fixed the problem!
Indeed, it was not a memory leak or corruption, but the library was indeed corrupted!

The problem was: edb.so depends on libedbutil which we recently made to depend on libmysqld! edb.so was loaded in the MySQL servers and libmysqld would overwrite symbols from the server. The problem is that the server and the embedded server were different versions!

I fixed the problem by moving the few functions that used sql interface from util in a file in edb/sql_utils so util no longer depends on libmysqld. I may have been able to fix it by installing a server of the same version, but I anyways did not like that utils depended on libmysqld for only two minor functions.